segment marquee, leave backward compat
