+++ lib/Marquee/Item.js
@@ -1,0 +1,486 @@
+require('moonstone');
+
+/**
+* Contains the declaration for the {@link moon.MarqueeSupport} mixin and the {@link moon.MarqueeText} &
+* {@link moon.MarqueeDecorator} kinds.
+* @module moonstone/Marquee
+*/
+
+var 
+	kind = require('enyo/kind'),
+	dom = require('enyo/dom'),
+	platform = require('enyo/platform'),
+	util = require('enyo/utils'),
+	Control = require('enyo/Control');
+
+var
+	options = require('../options'),
+	HighlightText = require('../HighlightText');
+
+module.exports = {
+
+	/**
+	* @private
+	*/
+	events: {
+
+		/**
+		* {@link moon.MarqueeItem#onMarqueeEnded}
+		*/
+		onMarqueeEnded:''
+	},
+
+	/**
+	* @private
+	*/
+	_marqueeItem_Handlers: {
+		onRequestMarquee: '_marquee_requestMarquee',
+		onRequestMarqueeStart: '_marquee_startAnimation',
+		onRequestMarqueeStop: '_marquee_stopAnimation',
+		onRequestMarqueeEnable: '_marquee_enable',
+		onRequestMarqueeDisable: '_marquee_disable',
+		ontransitionend: '_marquee_animationEnded'
+	},
+
+	/**
+	* @private
+	*/
+	observers: {
+		_marquee_contentChanged: ['content'],
+		_marquee_centeredChanged: ['centered'],
+		_marquee_wrapInsteadOfMarqueeChanged: ['wrapInsteadOfMarquee']
+	},
+
+	/**
+	* @private
+	*/
+	bindings: [
+		{from: '.allowHtml', to:'.$.marqueeText.allowHtml'}
+	],
+
+	/**
+	* @private
+	*/
+	classes: 'moon-marquee',
+
+	/**
+	* @method
+	* @private
+	*/
+	dispatchEvent: kind.inherit(function (sup) {
+		return function (sEventName, oEvent, oSender) {
+			if (sup.apply(this, arguments)) {
+				return true;
+			}
+			if (oEvent && !oEvent.delegate) {
+				var handler = this._marqueeItem_Handlers[sEventName];
+				if (handler && this[handler](oSender, oEvent)) {
+					return true;
+				}
+			}
+		};
+	}),
+
+	/**
+	* @private
+	*/
+	_marquee_enabled: true,
+
+	/**
+	* @private
+	*/
+	_marquee_distance: null,
+
+	/**
+	* @private
+	*/
+	_marquee_fits: null,
+
+	/**
+	* @private
+	*/
+	_marquee_puppetMaster: null,
+
+	/**
+	* @method
+	* @private
+	*/
+	create: kind.inherit(function (sup) {
+		return function () {
+			sup.apply(this, arguments);
+			this.detectTextDirectionality();
+			this._marquee_wrapInsteadOfMarqueeChanged();
+		};
+	}),
+
+	/**
+	* @method
+	* @private
+	*/
+	rendered: kind.inherit(function (sup) {
+		return function () {
+			sup.apply(this, arguments);
+			// There is a known issue where a parent control that modifies the layout will
+			// invalidate the measurements used to detect the proper alignment, which can
+			// result in the appropriate text-align rule not being applied. For example, this
+			// can occur with a moon.Header that is located inside a moon.Scroller which has
+			// vertical scrollbars visible.
+			this._marquee_detectAlignment();
+			setTimeout(util.bindSafely(this, this._marquee_calcDistance), platform.firefox ? 100 : 16);
+		};
+	}),
+
+	/**
+	* @method
+	* @private
+	*/
+	reflow: kind.inherit(function (sup) {
+		return function () {
+			sup.apply(this, arguments);
+			this._marquee_invalidateMetrics();
+			this._marquee_calcDistance();
+		};
+	}),
+
+	/**
+	* @method
+	* @private
+	*/
+	showingChangedHandler: kind.inherit(function (sup) {
+		return function () {
+			sup.apply(this, arguments);
+			this._marquee_reset();
+		};
+	}),
+
+	/**
+	* We must measure the content (after render) to determine if it's marqueeable, then to set
+	* its alignment to left if the content was explicitly set to LTR earlier. This happens when
+	* the locale is set to a RTL language, but your string contains no RTL characters in it.
+	* Therefore it's LTR, and if it's marqueeable, should be left aligned, so it marquees in the
+	* natural marqueeing direction.
+	*
+	* @param {Boolean} [forceAnimate]  Override the animation check (only accepts `true`). Use
+	*	this if you know already, because you've already measured that you will need to marquee.
+	* @param {Boolean} [forceRtl]  Override the internal RTL property, in case you know better.
+	* @private
+	*/
+	_marquee_detectAlignment: function (forceAnimate, forceRtl) {
+		var alignment = null,
+			rtl = forceRtl || this.rtl;
+
+		// We only attempt to set the alignment of this control if the locale's directionality
+		// differs from the directionality of our current marqueeable control (as determined by
+		// the control's content or is explicitly specified).
+		if (Control.prototype.rtl != rtl || this.centered) {
+			// If we will be marqueeing, we know the alignment needs to be set based on directionality.
+			if (forceAnimate || this._marquee_shouldAnimate()) {
+				if (rtl) {
+					alignment = 'right';
+				} else {
+					alignment = 'left';
+				}
+			}
+			// Alignment wasn't set yet, so we know we don't need to animate. Now we can center the text if we're supposed to.
+			if (!alignment && this.centered) {
+				alignment = 'center';
+			}
+		}
+
+		this.set('_marquee_alignment', alignment);
+	},
+
+	/**
+	* Reset the marquee distance if the alignment changes, since now we'll have to calculate the
+	* size again.
+	*
+	* @private
+	*/
+	_marquee_alignmentChanged: function () {
+		this.applyStyle('text-align', this._marquee_alignment);
+		this._marquee_invalidateMetrics();
+	},
+
+	/**
+	* @private
+	*/
+	_marquee_invalidateMetrics: function () {
+		this._marquee_distance = null;
+		this._marquee_fits = null;
+	},
+
+	/**
+	* When the content of this control changes, updates the content of
+	* `this.$.marqueeText` (if it exists).
+	*
+	* @private
+	*/
+	_marquee_contentChanged: function () {
+		this.detectTextDirectionality();
+		if (this.$.marqueeText) {
+			this.$.marqueeText.setContent(this.content);
+		}
+		if (this.generated) {
+			this._marquee_invalidateMetrics();
+			this._marquee_detectAlignment();
+			if (this.getAbsoluteShowing()) {
+				this._marquee_calcDistance();
+			}
+		}
+		this._marquee_reset();
+	},
+
+	/**
+	* If this control needs to marquee, lets the event originator know.
+	*
+	* @private
+	*/
+	_marquee_requestMarquee: function (sender, ev) {
+		if (!ev || !this.showing || this._marquee_fits) {
+			return;
+		}
+
+		this._marquee_puppetMaster = ev.originator;
+		ev.originator.addMarqueeItem(this);
+
+		this.marqueePause = ev.marqueePause || 1000;
+		this.marqueeSpeed = ev.marqueeSpeed || 60;
+	},
+
+	/**
+	* Starts marquee animation.
+	*
+	* @private
+	*/
+	_marquee_startAnimation: function (sender, ev) {
+		// if this control hasn't been generated, there's no need to follow through on
+		// marquee requests as we'll be unable to correctly measure the distance delta yet
+		if (!this.generated) return;
+
+		var distance = this._marquee_calcDistance();
+
+		// If there is no need to animate, return early
+		if (!this._marquee_shouldAnimate(distance)) {
+			this._marquee_fits = true;
+			this.doMarqueeEnded();
+			return;
+		}
+
+		// Lazy creation of _this.$.marqueeText_
+		if (!this.$.marqueeText) {
+			this._marquee_createMarquee();
+			distance = this._marquee_calcDistance();
+		}
+
+		this._marquee_addAnimationStyles(distance);
+
+		if (this.$.marqueeText) { return true; }
+	},
+
+	/**
+	* @private
+	*/
+	_marquee_enable: function () {
+		this.set('_marquee_enabled', true);
+	},
+
+	/**
+	* @private
+	*/
+	_marquee_disable: function () {
+		this.set('_marquee_enabled', false);
+		this._marquee_stopAnimation();
+	},
+
+	/**
+	* Stops marquee animation.
+	*
+	* @fires moon.MarqueeItem#onMarqueeEnded
+	* @private
+	*/
+	_marquee_stopAnimation: function (sender, ev) {
+		this.stopJob('stopMarquee');
+		this._marquee_removeAnimationStyles();
+		this.doMarqueeEnded();
+	},
+
+	/**
+	* When animation ends, starts `this.stopMarquee` job.
+	*
+	* @private
+	*/
+	_marquee_animationEnded: function (sender, ev) {
+		if (ev.originator !== this.$.marqueeText) {
+			return;
+		}
+
+		this.startJob('stopMarquee', '_marquee_stopAnimation', this.marqueePause);
+		return true;
+	},
+
+	/**
+	* Returns `true` if this control has enough content to animate.
+	*
+	* @private
+	*/
+	_marquee_shouldAnimate: function (distance) {
+		distance = (distance && distance >= 0) ? distance : this._marquee_calcDistance();
+		return (distance > 0);
+	},
+
+	/**
+	* Determines how far the marquee needs to scroll.
+	*
+	* @private
+	*/
+	_marquee_calcDistance: function () {
+		var node = this.$.marqueeText ? this.$.marqueeText.hasNode() : this.hasNode(),
+			rect;
+
+		if (node && this._marquee_distance == null) {
+			rect = node.getBoundingClientRect();
+			this._marquee_distance = Math.floor(Math.abs(node.scrollWidth - rect.width));
+
+			//if the distance is exactly 0, then the ellipsis
+			//most likely are hiding the content, and marquee does not
+			//need to animate
+			if(this._marquee_distance === 0) {
+				this.applyStyle('text-overflow', 'clip');
+			} else {
+				this.applyStyle('text-overflow', 'ellipsis');
+			}
+		}
+
+		return this._marquee_distance;
+	},
+
+	/**
+	* Returns duration based on `distance` and `this.marqueeSpeed`.
+	*
+	* @private
+	*/
+	_marquee_calcDuration: function (distance) {
+		return distance / this.marqueeSpeed;
+	},
+
+	/**
+	* Creates a marquee-able `div` inside of `this`.
+	*
+	* @private
+	*/
+	_marquee_createMarquee: function () {
+		// Do not create marqueeText when there is components block,
+		// because we don't know what should be the controlParent
+		if (this.components && this.components.length > 0) return;
+		var marqueeText = {name: 'marqueeText', kind: Control, classes: 'moon-marquee-text', allowHtml: this.allowHtml, content: this.content},
+			highlightText = null;
+
+		if (this instanceof HighlightText) {
+			dom.setInnerHtml(this.hasNode(), '');
+			highlightText = {renderDelegate: this.renderDelegate, highlightClasses: this.highlightClasses, search: this.search};
+			marqueeText = util.mixin(marqueeText, highlightText);
+		}
+		this.createComponent({name: 'marqueeTextWrapper', kind: Control, classes: 'moon-marquee-text-wrapper', components: [marqueeText]});
+		this.render();
+		return true;
+	},
+
+	/**
+	* @private
+	*/
+	_marquee_addAnimationStyles: function (distance) {
+		if (!this.$.marqueeText) return;
+		var duration = this._marquee_calcDuration(distance);
+
+		this.$.marqueeText.addClass('animate-marquee');
+
+		if (options.accelerate) {
+			dom.transform(this.$.marqueeText, {translateZ: 0});
+			this.$.marqueeText.applyStyle('transition', 'transform ' + duration + 's linear');
+			this.$.marqueeText.applyStyle('-webkit-transition', '-webkit-transform ' + duration + 's linear');
+		} else {
+			this.$.marqueeText.applyStyle('transition', 'left ' + duration + 's linear');
+			this.$.marqueeText.applyStyle('-webkit-transition', 'left ' + duration + 's linear');
+		}
+
+		// Need this timeout for FF!
+		setTimeout(this.bindSafely(function () {
+			if (options.accelerate) {
+				dom.transform(this.$.marqueeText, {translateX: this._marquee_adjustDistanceForRTL(distance) + 'px'});
+			} else {
+				this.$.marqueeText.applyStyle('left', this._marquee_adjustDistanceForRTL(distance) + 'px');
+			}
+		}), platform.firefox ? 100 : 16);
+	},
+
+	/**
+	* @private
+	*/
+	_marquee_removeAnimationStyles: function () {
+		if (!this.$.marqueeText) {
+			return;
+		}
+
+		this.$.marqueeText.applyStyle('transition-duration', '0s');
+		this.$.marqueeText.applyStyle('-webkit-transition-duration', '0s');
+
+		// Need this timeout for FF!
+		/**
+		* @private
+		*/
+		setTimeout(this.bindSafely(function () {
+			this.$.marqueeText.removeClass('animate-marquee');
+			if (options.accelerate) {
+				dom.transform(this.$.marqueeText, {translateX: null, translateZ: null});
+			} else {
+				this.$.marqueeText.applyStyle('left', null);
+			}
+		}), platform.firefox ? 100 : 0);
+	},
+
+	/**
+	* Flips distance value for RTL support.
+	*
+	* @private
+	*/
+	_marquee_adjustDistanceForRTL: function (distance) {
+		return this.rtl ? distance : distance * -1;
+	},
+
+	/**
+	* @private
+	*/
+	_marquee_reset: function () {
+		this._marquee_invalidateMetrics();
+		if (this._marquee_puppetMaster) {
+			this._marquee_puppetMaster.resetMarquee();
+		}
+	},
+
+	/**
+	* @private
+	*/
+	_marquee_centeredChanged: function () {
+		this._marquee_detectAlignment();
+	},
+
+	/**
+	* @private
+	*/
+	_marquee_wrapInsteadOfMarqueeChanged: function(old) {
+		if (this.wrapInsteadOfMarquee) {
+			this.addClass('allow-wrap');
+			if (this.$.marqueeText) {
+				this.$.marqueeTextWrapper.destroy();
+				this.render();
+			}
+		}
+		if (old && !this.wrapInsteadOfMarquee) {
+			this.removeClass('allow-wrap');
+			// FIXME: Performing creation here to workaround potential WebKit measuring issue
+			// with scrollWidth (under-measures by 10px when marquee components are destroyed
+			// when we switch wrapInsteadofMarquee from `false` to `true`, and back to `false`).
+			this._marquee_createMarquee();
+		}
+	}
+};
